checkpoint-5/src/Extra_Local_src_files/CP5_Q1_Q3/requirements.txt
sklearn
nltk 
pandas
matplotlib
psycopg2-binary